Reorganize token components
