--- src/modules/admin/components/Tokens/Tokens.css
@@ -1,43 +1,0 @@
-.main {
-  display: grid;
-  padding: var(--pad-huge) 0 0 8px;
-  height: 100%;
-  grid-template-columns: minmax(700px, auto) 180px;
-  grid-template-rows: 1fr;
-  column-gap: 80px;
-}
-
-.main main {
-  display: flex;
-  flex-direction: column;
-}
-
-.titleContainer {
-  display: flex;
-  justify-content: space-between;
-  align-items: baseline;
-  border-bottom: 1px solid var(--temp-grey-5);
-
-  & h3 {
-    /*
-     * @NOTE We can't compose from `inlineEllipsis` here since we're inside an
-     * element, not a component :(
-     */
-    width: calc(100% - 150px);
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-}
-
-.sidebar {
-  margin-top: 80px;
-}
-
-.sidebar li {
-  padding: 10px 0;
-}
-
-.mainContent {
-  flex: 1;
-}
