--- src/modules/admin/components/Tokens/Tokens.css.d.ts
@@ -1,4 +1,0 @@
-export const main: string;
-export const titleContainer: string;
-export const sidebar: string;
-export const mainContent: string;
