+++ src/modules/dashboard/components/ColonyFunding/ColonyFunding.css
@@ -1,16 +1,48 @@
 
 .main {
   display: flex;
-  flex-direction: column;
-  padding: 22px 0 20px 20px;
-  min-height: 100%;
+  display: grid;
+  padding: 22px 0 0 22px;
+
+  /* @fixme use correct height / growth here */
+  height: calc(100% - 100px);
+
+  /* @fixme find better way to set these widths - select wallet on this screen to see it change */
+  grid-template-columns: auto minmax(48%, 700px) auto;
 }
 
-.backLink {
+.titleContainer {
   display: flex;
-  align-items: center;
+  justify-content: space-between;
+  align-items: baseline;
+  margin-bottom: 35px;
+  border-bottom: 1px solid var(--temp-grey-5);
+
+  & h3 {
+    /*
+     * @NOTE We can't compose from `inlineEllipsis` here since we're inside an
+     * element, not a component :(
+     */
+    width: calc(100% - 150px);
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+  }
 }
 
 .content {
-  padding-left: 21.25%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+  min-height: 0;
+  grid-column-start: 2;
+}
+
+.banner {
+  margin-top: 35px;
+}
+
+.aside {
+  padding: 75px 36%;
 }
