+++ src/modules/dashboard/components/ColonyFunding/ColonyFunding.css.d.ts
@@ -1,3 +1,5 @@
 export const main: string;
-export const backLink: string;
+export const titleContainer: string;
 export const content: string;
+export const banner: string;
+export const aside: string;
