+++ src/modules/dashboard/components/ColonyFunding/ColonyFunding.tsx
@@ -1,10 +1,28 @@
-import React from 'react';
+import React, { ComponentProps, useMemo, useState } from 'react';
+import { defineMessages, useIntl } from 'react-intl';
 import { RouteChildrenProps } from 'react-router';
+import sortBy from 'lodash/sortBy';
 
-import Tokens from '~admin/Tokens';
+import { COLONY_TOTAL_BALANCE_DOMAIN_ID } from '~constants';
+import { Form, Select } from '~core/Fields';
+import Heading from '~core/Heading';
+import ColonyFundingBanner from '~dashboard/ColonyFundingBanner';
+import ColonyFundingMenu from '~dashboard/ColonyFundingMenu';
+import TokenCardList from '~dashboard/TokenCardList';
 import { useColonyFromNameQuery } from '~data/index';
 
 import styles from './ColonyFunding.css';
+
+const MSG = defineMessages({
+  labelSelectDomain: {
+    id: 'dashboard.ColonyFunding.labelSelectDomain',
+    defaultMessage: 'Select a domain',
+  },
+  title: {
+    defaultMessage: 'Funds',
+    id: 'dashboard.ColonyFunding.title',
+  },
+});
 
 type Props = RouteChildrenProps<{ colonyName: string }>;
 
@@ -17,11 +35,50 @@ const ColonyFunding = ({ match }: Props) => {
     );
   }
   const { colonyName } = match.params;
+  const { formatMessage } = useIntl();
+
+  const [selectedDomainId, setSelectedDomainId] = useState<number>(
+    COLONY_TOTAL_BALANCE_DOMAIN_ID,
+  );
 
   const { data } = useColonyFromNameQuery({
     // We have to define an empty address here for type safety, will be replaced by the query
     variables: { name: colonyName, address: '' },
   });
+
+  const domainChoices = useMemo<
+    ComponentProps<typeof Select>['options']
+  >(() => {
+    if (!data || !data.colony) {
+      return [];
+    }
+    const {
+      colony: { domains },
+    } = data;
+    return [
+      {
+        value: COLONY_TOTAL_BALANCE_DOMAIN_ID.toString(),
+        label: { id: 'domain.all' },
+      },
+      ...sortBy(
+        Object.entries(domains || {})
+          .sort()
+          .map(([domainId, { name }]) => ({
+            label: name,
+            value: domainId.toString(),
+          })),
+        ['value'],
+      ),
+    ];
+  }, [data]);
+
+  const selectedDomainLabel: string = useMemo(() => {
+    const { label = '' } =
+      domainChoices.find(
+        ({ value }) => value === selectedDomainId.toString(),
+      ) || {};
+    return typeof label === 'string' ? label : formatMessage(label);
+  }, [domainChoices, formatMessage, selectedDomainId]);
 
   if (!data) {
     return null;
@@ -32,8 +89,49 @@ const ColonyFunding = ({ match }: Props) => {
   return (
     <div className={styles.main}>
       <div className={styles.content}>
-        <Tokens colony={colony} />
+        <div>
+          <div className={styles.titleContainer}>
+            <Heading
+              text={MSG.title}
+              textValues={{ selectedDomainLabel }}
+              appearance={{ size: 'medium', theme: 'dark' }}
+            />
+            <Form
+              initialValues={{
+                selectDomain: COLONY_TOTAL_BALANCE_DOMAIN_ID.toString(),
+              }}
+              onSubmit={() => {}}
+            >
+              <Select
+                appearance={{
+                  alignOptions: 'right',
+                  width: 'strict',
+                  theme: 'alt',
+                }}
+                elementOnly
+                label={MSG.labelSelectDomain}
+                name="selectDomain"
+                onChange={(value) => setSelectedDomainId(Number(value))}
+                options={domainChoices}
+              />
+            </Form>
+          </div>
+          <TokenCardList
+            appearance={{ numCols: '3' }}
+            nativeTokenAddress={colony.nativeTokenAddress}
+            tokens={colony.tokens}
+          />
+        </div>
+        <div className={styles.banner}>
+          <ColonyFundingBanner colonyAddress={colony.colonyAddress} />
+        </div>
       </div>
+      <aside className={styles.aside}>
+        <ColonyFundingMenu
+          selectedDomainId={selectedDomainId}
+          colony={colony}
+        />
+      </aside>
     </div>
   );
 };
