+++ src/modules/dashboard/components/ColonyFundingBanner/ColonyFundingBanner.css.d.ts
