+++ src/modules/dashboard/components/ColonyFundingBanner/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ColonyFundingBanner';
