+++ src/modules/dashboard/components/ColonyFundingMenu/ColonyFundingMenu.css
@@ -1,0 +1,3 @@
+.main li + li {
+  margin: 18px 0 0;
+}
