+++ src/modules/dashboard/components/ColonyFundingMenu/ColonyFundingMenu.css.d.ts
@@ -1,0 +1,1 @@
+export const main: string;
