+++ src/modules/dashboard/components/ColonyFundingMenu/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ColonyFundingMenu';
