+++ src/modules/dashboard/components/TokenCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TokenCard';
