+++ src/modules/dashboard/components/TokenCardList/TokenCardList.css
@@ -1,6 +1,5 @@
 
 .tokenCardContainer {
-  margin: 30px 0;
   padding-bottom: 10px;
   max-height: 510px;
   overflow: auto;
