+++ src/modules/dashboard/components/TokenCardList/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TokenCardList';
